backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.4.2
supabase==2.3.1
python-dotenv==1.0.1
numpy==1.25.2
pandas==2.1.1
numpy-financial==1.0.0
requests==2.31.0
cachetools==5.3.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
psutil==5.9.8